Refactor data processing services into a separate module
This commit improves the structure and modularity of the codebase by extracting data processing services from `gui/main.py` into their own dedicated module.

The following changes are included:

- **New Services Module:**
  - A new `services/` directory and `services/data_processing.py` file have been created.
  - The `services` directory has been made a Python package.

- **Move Service Functions:**
  - The `run_preprocessing_preview` and `generate_mesh_from_params` functions have been moved from `gui/main.py` to `services/data_processing.py`.

- **Update Imports:**
  - `gui/main.py` has been updated to import and expose the service functions from their new location.
  - `tests/test_gui_integration.py` has been updated to import from the new module.

This refactoring makes the responsibilities of `gui/main.py` clearer and improves the overall modularity of the backend code.
